requirements.txt
fastapi
uvicorn[standard]
passlib[bcrypt]
argon2-cffi
sqlalchemy>=2.0
psycopg[binary]
pydantic>=2
scikit-learn
sentence-transformers
python-dotenv
pypdf
openai>=1.55.0
pillow
pytesseract
pdf2image
